Turn the JwtTokenBuilder class into an API interface
